Add ExemptIPAddresses setting
Adds the `ExemptIPAddresses` setting to Settings.json. This allows the
ability to set a custom amount of connections per IP address.

The intention is for the default maximum amount of connections to
be handled with `MaxNonExempt` but to allow for extra connections to
account for houses with multiple people playing with the same IP address.